+++ dipy/viz/utils.py
@@ -14,6 +14,37 @@ vtk, have_vtk, setup_module = optional_package('vtk')
 ns, have_numpy_support, _ = optional_package('vtk.util.numpy_support')
 
 
+def set_input(vtk_object, inp):
+    """ Generic input function which takes into account VTK 5 or 6
+
+    Parameters
+    ----------
+    vtk_object: vtk object
+    inp: vtkPolyData or vtkImageData or vtkAlgorithmOutput
+
+    Returns
+    -------
+    vtk_object
+
+    Notes
+    -------
+    This can be used in the following way::
+        from dipy.viz.utils import set_input
+        poly_mapper = set_input(vtk.vtkPolyDataMapper(), poly_data)
+    """
+    if isinstance(inp, vtk.vtkPolyData) \
+       or isinstance(inp, vtk.vtkImageData):
+        if vtk.VTK_MAJOR_VERSION <= 5:
+            vtk_object.SetInput(inp)
+        else:
+            vtk_object.SetInputData(inp)
+    elif isinstance(inp, vtk.vtkAlgorithmOutput):
+        vtk_object.SetInputConnection(inp)
+
+    vtk_object.Update()
+    return vtk_object
+
+
 def numpy_to_vtk_points(points):
     """ Numpy points array to a vtk points array
 
@@ -55,37 +86,6 @@ def numpy_to_vtk_colors(colors):
     vtk_colors = ns.numpy_to_vtk(np.asarray(colors), deep=True,
                                  array_type=vtk.VTK_UNSIGNED_CHAR)
     return vtk_colors
-
-
-def set_input(vtk_object, inp):
-    """ Generic input function which takes into account VTK 5 or 6
-
-    Parameters
-    ----------
-    vtk_object: vtk object
-    inp: vtkPolyData or vtkImageData or vtkAlgorithmOutput
-
-    Returns
-    -------
-    vtk_object
-
-    Notes
-    -------
-    This can be used in the following way::
-        from dipy.viz.utils import set_input
-        poly_mapper = set_input(vtk.vtkPolyDataMapper(), poly_data)
-    """
-    if isinstance(inp, vtk.vtkPolyData) \
-       or isinstance(inp, vtk.vtkImageData):
-        if vtk.VTK_MAJOR_VERSION <= 5:
-            vtk_object.SetInput(inp)
-        else:
-            vtk_object.SetInputData(inp)
-    elif isinstance(inp, vtk.vtkAlgorithmOutput):
-        vtk_object.SetInputConnection(inp)
-
-    vtk_object.Update()
-    return vtk_object
 
 
 def map_coordinates_3d_4d(input_array, indices):
@@ -228,3 +228,198 @@ def lines_to_vtk_polydata(lines, colors=None):
     poly_data.SetLines(vtk_lines)
     poly_data.GetPointData().SetScalars(vtk_colors)
     return poly_data, is_colormap
+
+
+##########################################
+# Get PolyData properties in Numpy array
+##########################################
+def get_polydata_lines(line_polydata):
+    """ vtk polydata to a list of lines ndarrays
+
+    Parameters
+    ----------
+    line_polydata : vtkPolyData
+    
+    
+    Returns
+    -------
+    lines : list of N curves represented as 2D ndarrays
+    """
+    lines_vertices = ns.vtk_to_numpy(line_polydata.GetPoints().GetData())
+    lines_idx = ns.vtk_to_numpy(line_polydata.GetLines().GetData())
+    
+    lines = []
+    current_idx = 0
+    while current_idx < len(lines_idx):
+        line_len = lines_idx[current_idx]
+        #print line_len
+        next_idx = current_idx + line_len + 1 
+        line_range = lines_idx[current_idx + 1: next_idx]
+        #print line_range
+        lines += [lines_vertices[line_range]]
+        current_idx = next_idx
+    return lines
+
+
+def get_polydata_triangles(polydata):
+    """ get triangles (ndarrays Nx3 int) from a vtk polydata
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    
+    
+    Returns
+    -------
+    output : triangles, represented as 2D ndarrays (Nx3)
+    """
+    vtk_polys = ns.vtk_to_numpy(polydata.GetPolys().GetData())
+    assert((vtk_polys[::4] == 3).all())  # test if its really triangles
+    return np.vstack([vtk_polys[1::4], vtk_polys[2::4], vtk_polys[3::4]]).T
+
+def get_polydata_vertices(polydata):
+    """ get vertices (ndarrays Nx3 int) from a vtk polydata
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    
+    
+    Returns
+    -------
+    output : points, represented as 2D ndarrays (Nx3)
+    """
+    return ns.vtk_to_numpy(polydata.GetPoints().GetData())
+
+
+def get_polydata_normals(polydata):
+    """ get vertices normal (ndarrays Nx3 int) from a vtk polydata
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    
+    
+    Returns
+    -------
+    output : normals, represented as 2D ndarrays (Nx3)
+             None (if no normals in the vtk polydata)
+    """
+    vtk_normals = polydata.GetPointData().GetNormals()
+    if vtk_normals is None:
+        return None
+    else:
+        return ns.vtk_to_numpy(vtk_normals)
+
+
+def get_polydata_colors(polydata):
+    """ get points color (ndarrays Nx3 int) from a vtk polydata
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    
+    
+    Returns
+    -------
+    output : colors, represented as 2D ndarrays (Nx3)
+             None (if no normals in the vtk polydata)
+    """
+    vtk_colors = polydata.GetPointData().GetScalars()
+    if vtk_colors is None:
+        return None
+    else:
+        return ns.vtk_to_numpy(vtk_colors)
+
+
+##########################################
+# Set PolyData properties with Numpy array
+##########################################
+def set_polydata_triangles(polydata, triangles):
+    """ set polydata triangles with a numpy array (ndarrays Nx3 int) 
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    triangles : triangles, represented as 2D ndarrays (Nx3)
+    """
+    vtk_triangles = np.hstack(np.c_[np.ones(len(triangles)).astype(np.int) * 3, triangles])
+    vtk_triangles = ns.numpy_to_vtkIdTypeArray(vtk_triangles, deep=True)
+    vtk_cells = vtk.vtkCellArray()
+    vtk_cells.SetCells(len(triangles), vtk_triangles)
+    polydata.SetPolys(vtk_cells)
+    return polydata
+
+def set_polydata_vertices(polydata, vertices):
+    """ set polydata vertices with a numpy array (ndarrays Nx3 int) 
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    vertices : vertices, represented as 2D ndarrays (Nx3)
+    """
+    vtk_points = vtk.vtkPoints()
+    vtk_points.SetData(ns.numpy_to_vtk(vertices, deep=True))
+    polydata.SetPoints(vtk_points)
+    return polydata
+
+def set_polydata_normals(polydata, normals):
+    """ set polydata normals with a numpy array (ndarrays Nx3 int) 
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    normals : normals, represented as 2D ndarrays (Nx3) (one per vertex)
+    """
+    vtk_normals = ns.numpy_to_vtk(normals, deep=True)
+    polydata.GetPointData().SetNormals(vtk_normals)
+    return polydata
+
+def set_polydata_colors(polydata, colors):
+    """ set polydata colors with a numpy array (ndarrays Nx3 int) 
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    colors : colors, represented as 2D ndarrays (Nx3) 
+                colors are uint8 [0,255] RGB for each points
+    """
+    vtk_colors = ns.numpy_to_vtk(colors, deep=True, array_type=vtk.VTK_UNSIGNED_CHAR)
+    vtk_colors.SetNumberOfComponents(3)
+    vtk_colors.SetName("RGB")
+    polydata.GetPointData().SetScalars(vtk_colors)
+    return polydata
+
+
+##########################################
+# Update/ Refresh PolyData
+##########################################
+def update_polydata_normals(polydata):
+    """ generate and update polydata normal
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    """
+    normals_gen = set_input(vtk.vtkPolyDataNormals(), polydata)
+    normals_gen.ComputePointNormalsOn()
+    normals_gen.ComputeCellNormalsOn()
+    normals_gen.SplittingOff()
+    #normals_gen.FlipNormalsOn()
+    #normals_gen.ConsistencyOn()
+    #normals_gen.AutoOrientNormalsOn()
+    normals_gen.Update()
+
+    vtk_normals = normals_gen.GetOutput().GetPointData().GetNormals()
+    polydata.GetPointData().SetNormals(vtk_normals)
+
+def update_polydata(polydata):
+    """ update polydata
+
+    Parameters
+    ----------
+    polydata : vtkPolyData
+    """
+    if vtk.VTK_MAJOR_VERSION <= 5:
+        polydata.Update()
+
